+++ packages/rockets-next-web/src/app/showroom/users/types.ts
@@ -1,0 +1,7 @@
+export interface FormData {
+  id?: string;
+  email: string;
+  username: string;
+}
+
+export type ActionType = "creation" | "edit" | "details" | "delete" | null;
